+++ src/tiled.frame.cpp
@@ -1,4 +1,5 @@
 #include "tiled.frame.hh"
+#include "zarr.hh"
 
 #include <algorithm>
 #include <cstring>
@@ -40,6 +41,8 @@ TiledFrame::TiledFrame(const VideoFrame* frame,
                        const ImageShape& image_shape,
                        const TileShape& tile_shape)
   : bytes_of_image_{ 0 }
+  , frame_id_{ 0 }
+  , layer_{ 0 }
   , image_shape_{ image_shape }
   , tile_shape_{ tile_shape }
 {
@@ -55,6 +58,62 @@ TiledFrame::TiledFrame(const VideoFrame* frame,
     frame_id_ = frame->frame_id;
 }
 
+TiledFrame::TiledFrame(uint64_t frame_id,
+                       size_t layer,
+                       const ImageShape& image_shape,
+                       const TileShape& tile_shape)
+  : bytes_of_image_{ get_bytes_per_frame(image_shape) }
+  , frame_id_{ frame_id }
+  , layer_{ layer }
+  , image_shape_{ image_shape }
+  , tile_shape_{ tile_shape }
+{
+    buf_.resize(bytes_of_image_);
+    std::fill_n(buf_.begin(), bytes_of_image_, 0);
+}
+
+size_t
+TiledFrame::bytes_of_image() const
+{
+    return bytes_of_image_;
+}
+
+const ImageShape&
+TiledFrame::image_shape() const
+{
+    return image_shape_;
+}
+
+const TileShape&
+TiledFrame::tile_shape() const
+{
+    return tile_shape_;
+}
+
+uint64_t
+TiledFrame::frame_id() const
+{
+    return frame_id_;
+}
+
+size_t
+TiledFrame::layer() const
+{
+    return layer_;
+}
+
+const uint8_t*
+TiledFrame::image() const
+{
+    return buf_.data();
+}
+
+uint8_t*
+TiledFrame::data()
+{
+    return buf_.data();
+}
+
 size_t
 TiledFrame::copy_tile(uint8_t* tile,
                       size_t bytes_of_tile,
@@ -63,7 +122,7 @@ TiledFrame::copy_tile(uint8_t* tile,
                       uint32_t tile_plane) const
 {
     CHECK(tile);
-    CHECK(bytes_of_tile == bytes_per_tile(image_shape_, tile_shape_));
+    CHECK(bytes_of_tile >= bytes_per_tile(image_shape_, tile_shape_));
     memset(tile, 0, bytes_of_tile);
 
     uint8_t* region = nullptr;