+++ src/tiled.frame.hh
@@ -26,8 +26,21 @@ class TiledFrame
     TiledFrame(const VideoFrame* frame,
                const ImageShape&,
                const TileShape& tile_shape);
+    TiledFrame(uint64_t frame_id,
+               size_t layer,
+               const ImageShape& image_shape,
+               const TileShape& tile_shape);
     TiledFrame(const TiledFrame&) = delete;
     ~TiledFrame() = default;
+
+    size_t bytes_of_image() const;
+    const ImageShape& image_shape() const;
+    const TileShape& tile_shape() const;
+
+    uint64_t frame_id() const;
+    size_t layer() const;
+    const uint8_t* image() const;
+    uint8_t* data();
 
     /// @brief Copy the tile indexed by @p tile_col, @p tile_row, and
     ///        @p tile_plane into the buffer at @p tile.
@@ -47,6 +60,7 @@ class TiledFrame
   private:
     size_t bytes_of_image_;
     uint64_t frame_id_;
+    size_t layer_;
     std::vector<uint8_t> buf_;
     ImageShape image_shape_;
     TileShape tile_shape_;