+++ tests/write-zarr-with-zstd-compression.cpp
@@ -1,13 +1,17 @@
-#include "device/hal/device.manager.h"
 #include "acquire.h"
+#include "device/hal/device.manager.h"
 #include "platform.h"
 #include "logger.h"
 
 #include <cstring>
+#include <fstream>
 #include <filesystem>
 #include <stdexcept>
 
+#include "json.hpp"
+
 namespace fs = std::filesystem;
+using json = nlohmann::json;
 
 void
 reporter(int is_error,
@@ -46,8 +50,16 @@ reporter(int is_error,
 #define DEVOK(e) CHECK(Device_Ok == (e))
 #define OK(e) CHECK(AcquireStatus_Ok == (e))
 
+/// example: `ASSERT_EQ(int,"%d",42,meaning_of_life())`
+#define ASSERT_EQ(T, fmt, a, b)                                                \
+    do {                                                                       \
+        T a_ = (T)(a);                                                         \
+        T b_ = (T)(b);                                                         \
+        EXPECT(a_ == b_, "Expected %s==%s but " fmt "!=" fmt, #a, #b, a_, b_); \
+    } while (0)
+
 /// Check that a>b
-/// example: `ASSERT_GT(int,"%d",42,meaning_of_life())`
+/// example: `ASSERT_GT(int,"%d",43,meaning_of_life())`
 #define ASSERT_GT(T, fmt, a, b)                                                \
     do {                                                                       \
         T a_ = (T)(a);                                                         \
@@ -56,9 +68,9 @@ reporter(int is_error,
           a_ > b_, "Expected (%s) > (%s) but " fmt "<=" fmt, #a, #b, a_, b_);  \
     } while (0)
 
-static const size_t nframes = 70;
-static const size_t image_width = 64;
-static const size_t image_height = 48;
+static const uint32_t expected_frames_per_chunk = 71;
+static const uint32_t frame_width = 64;
+static const uint32_t frame_height = 48;
 
 void
 acquire(AcquireRuntime* runtime, const char* filename)
@@ -91,15 +103,15 @@ acquire(AcquireRuntime* runtime, const char* filename)
                                   sample_spacing_um));
 
     storage_properties_set_chunking_props(
-      &props.video[0].storage.settings, image_width, image_height, 1, 64 << 20);
+      &props.video[0].storage.settings, frame_width, frame_height, 1, 64 << 20);
 
     props.video[0].camera.settings.binning = 1;
     props.video[0].camera.settings.pixel_type = SampleType_u8;
-    props.video[0].camera.settings.shape = { .x = image_width,
-                                             .y = image_height };
+    props.video[0].camera.settings.shape = { .x = frame_width,
+                                             .y = frame_height };
     // we may drop frames with lower exposure
     props.video[0].camera.settings.exposure_time_us = 1e4;
-    props.video[0].max_frame_count = nframes;
+    props.video[0].max_frame_count = expected_frames_per_chunk;
 
     OK(acquire_configure(runtime, &props));
     OK(acquire_start(runtime));
@@ -113,25 +125,44 @@ main()
     acquire(runtime, TEST ".zarr");
 
     CHECK(fs::is_directory(TEST ".zarr"));
-    const auto data_file_path = fs::path(TEST ".zarr/0/0/0/0/0");
-    CHECK(fs::is_regular_file(data_file_path));
-
-    const auto raw_file_size = nframes * image_width * image_height;
-    ASSERT_GT(size_t, "%zu", fs::file_size(data_file_path), 0);
-    ASSERT_GT(size_t, "%zu", raw_file_size, fs::file_size(data_file_path));
-
-    const auto zarray_path = fs::path(TEST ".zarr") / "0" / ".zarray";
-    CHECK(fs::is_regular_file(zarray_path));
-    CHECK(fs::file_size(zarray_path) > 0);
 
     const auto external_metadata_path =
       fs::path(TEST ".zarr") / "0" / ".zattrs";
     CHECK(fs::is_regular_file(external_metadata_path));
-    CHECK(fs::file_size(external_metadata_path) > 0);
+    ASSERT_GT(int, "%d", fs::file_size(external_metadata_path), 0);
 
     const auto group_zattrs_path = fs::path(TEST ".zarr") / ".zattrs";
     CHECK(fs::is_regular_file(group_zattrs_path));
-    CHECK(fs::file_size(group_zattrs_path) > 0);
+    ASSERT_GT(int, "%d", fs::file_size(group_zattrs_path), 0);
+
+    const auto zarray_path = fs::path(TEST ".zarr") / "0" / ".zarray";
+    CHECK(fs::is_regular_file(zarray_path));
+    ASSERT_GT(int, "%d", fs::file_size(zarray_path), 0);
+
+    // check metadata
+    std::ifstream f(zarray_path);
+    json zarray = json::parse(f);
+
+    auto shape = zarray["shape"];
+    ASSERT_EQ(int, "%d", expected_frames_per_chunk, shape[0]);
+    ASSERT_EQ(int, "%d", 1, shape[1]);
+    ASSERT_EQ(int, "%d", frame_height, shape[2]);
+    ASSERT_EQ(int, "%d", frame_width, shape[3]);
+
+    auto chunks = zarray["chunks"];
+    ASSERT_EQ(int, "%d", expected_frames_per_chunk, chunks[0]);
+    ASSERT_EQ(int, "%d", 1, chunks[1]);
+    ASSERT_EQ(int, "%d", frame_height, chunks[2]);
+    ASSERT_EQ(int, "%d", frame_width, chunks[3]);
+
+    // check chunked data
+    auto raw_chunk_size = chunks[0].get<int>() * chunks[1].get<int>() *
+                          chunks[2].get<int>() * chunks[3].get<int>();
+
+    const auto chunk_file_path = fs::path(TEST ".zarr/0/0/0/0/0");
+    CHECK(fs::is_regular_file(chunk_file_path));
+    ASSERT_GT(int, "%d", fs::file_size(chunk_file_path), 0);
+    ASSERT_GT(int, "%d", raw_chunk_size, fs::file_size(chunk_file_path));
 
     LOG("Done (OK)");
     acquire_shutdown(runtime);